requirements.txt
bottle
simplejson
gevent
gevent_zeromq
nose
requests
python-dateutil<2.0
sphinx
pyzmq
pytz